Move to another module and use in flat index
