+++ crates/uv-client/src/lib.rs
@@ -7,6 +7,7 @@ pub use registry_client::{
 };
 pub use rkyvutil::OwnedArchive;
 
+mod auth;
 mod cached_client;
 mod error;
 mod flat_index;
